requirements.txt
MySQL-python==1.2.4
PyAudio==0.2.7
pyspotify==1.10
requests==1.2.3
tornado==3.0.2
torndb==0.1
wsgiref==0.1.2
